requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
SQLAlchemy==2.0.35
sqlparse==0.5.1
typing_extensions==4.12.2
